feat: add overlay ability to the PageLayout component
